1.解决登录等待问题 2.充电桩List界面，测试费用信息显示 3.添加LED控件，并设置LED位置，用于在充电桩List界面显示通信状态 4.设置通信状态数据，并完成测试 5.解决LED控件页面布局BUG -----------------下页--------------------- 1.设计故障记录数据库 2.实现故障记录的数据库数据存取 3.解决数据库存数据BUG，数据转换问题 4.重新定义部分与服务器交互接口

+++ chargePileManage/Classes/Common/Lib/DXPopover/DXPopover.m
@@ -71,7 +71,7 @@
         frame.origin.x -= (outerSideEdge+sideEdge);
     } else {
         if (CGRectGetMinX(frame)<0) {
-            frame.origin.x += fabsf(CGRectGetMinX(frame))+sideEdge;
+            frame.origin.x += fabs(CGRectGetMinX(frame))+sideEdge;
         }
     }
